fix: resolve SQLAlchemy BigInteger import and add time module
- Fixed SQLAlchemy column types (BigInteger instead of BigInt)
- Added missing time import in main.py
- Removed UI from docker-compose (runs locally for dev)
- Added .dockerignore for faster UI builds
- Created E2E test report documentation
- All 8 backend services running successfully

+++ dataquarantine/core/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, BigInt, Text
+from sqlalchemy import Column, Integer, String, DateTime, BigInteger, Text
 from sqlalchemy.dialects.postgresql import UUID
 import datetime
 import uuid
@@ -10,7 +10,7 @@ class QuarantineRecord(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     topic = Column(String, nullable=False)
     partition = Column(Integer, nullable=False)
-    offset = Column(BigInt, nullable=False)
+    offset = Column(BigInteger, nullable=False)
     timestamp = Column(DateTime, nullable=False)
     schema_name = Column(String, nullable=False)
     schema_version = Column(String, nullable=False)
